Split AppleWin -> CmdLine for the command line option parsing.
