+++ AppleWinExpress2019.vcxproj.filters
@@ -217,6 +217,9 @@
     <ClCompile Include="source\Windows\AppleWin.cpp">
       <Filter>Source Files\Windows</Filter>
     </ClCompile>
+    <ClCompile Include="source\CmdLine.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="source\CommonVICE\6510core.h">
@@ -518,6 +521,9 @@
     </ClInclude>
     <ClInclude Include="source\Windows\AppleWin.h">
       <Filter>Source Files\Windows</Filter>
+    </ClInclude>
+    <ClInclude Include="source\CmdLine.h">
+      <Filter>Source Files</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
