+++ source/Registry.cpp
@@ -29,8 +29,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "StdAfx.h"
 
 #include "Registry.h"
-
-extern std::string g_sConfigFile;
+#include "CmdLine.h"
 
 namespace _ini {
 	//===========================================================================
